requirements.txt
# Core dependencies
python-binance==1.0.17
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
python-dotenv==1.0.0
ccxt==4.1.22
ta==0.11.0
schedule==1.2.1
websocket-client==1.7.0
pyyaml==6.0.1
telegram-send==0.37
scipy==1.12.0

# API server dependencies
fastapi==0.110.0
uvicorn==0.27.1
websockets==12.0
python-multipart==0.0.9
pydantic==2.5.2
SQLAlchemy==2.0.27
jwt==1.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Dashboard dependencies
streamlit==1.32.0
plotly==5.18.0

# Testing and development
pytest==7.4.3
pytest-mock==3.12.0
jupyter==1.0.0

# Performance optimization
numba==0.58.1
joblib==1.3.2
